# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AtomQuest Goal Setting & Tracking Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users (admin/manager)
      parameters:
        - name: role
          in: query
          schema:
            type: string
        - name: department
          in: query
          schema:
            type: string
        - name: managerId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a new user (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get a user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete a user (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /goals:
    get:
      operationId: listGoals
      tags: [goals]
      summary: List goals (filtered by role context)
      parameters:
        - name: employeeId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
        - name: approvalStatus
          in: query
          schema:
            type: string
        - name: quarter
          in: query
          schema:
            type: string
        - name: department
          in: query
          schema:
            type: string
        - name: thrustArea
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of goals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Goal"
    post:
      operationId: createGoal
      tags: [goals]
      summary: Create a new goal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoalInput"
      responses:
        "201":
          description: Goal created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
  /goals/{id}:
    get:
      operationId: getGoal
      tags: [goals]
      summary: Get a goal by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Goal found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
        "404":
          description: Not found
    patch:
      operationId: updateGoal
      tags: [goals]
      summary: Update a goal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoalUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
    delete:
      operationId: deleteGoal
      tags: [goals]
      summary: Delete a draft goal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /shared-goals:
    get:
      operationId: listSharedGoals
      tags: [shared-goals]
      summary: List shared/departmental goals
      responses:
        "200":
          description: List of shared goals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SharedGoal"
    post:
      operationId: createSharedGoal
      tags: [shared-goals]
      summary: Create a shared goal (manager/admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SharedGoalInput"
      responses:
        "201":
          description: Shared goal created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SharedGoal"
  /shared-goals/{id}:
    get:
      operationId: getSharedGoal
      tags: [shared-goals]
      summary: Get a shared goal by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Shared goal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SharedGoal"
    patch:
      operationId: updateSharedGoal
      tags: [shared-goals]
      summary: Update a shared goal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SharedGoalUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SharedGoal"

  # ─── QUARTERLY CYCLES ───────────────────────────────────
  /quarters:
    get:
      operationId: listQuarters
      tags: [quarterly]
      summary: List all quarters and their active status
      responses:
        "200":
          description: List of quarters
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Quarter"
    post:
      operationId: createQuarter
      tags: [quarterly]
      summary: Create a new quarter (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuarterInput"
      responses:
        "201":
          description: Quarter created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quarter"
  /quarters/{id}:
    patch:
      operationId: updateQuarter
      tags: [quarterly]
      summary: Update quarter (admin — toggle active, override)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuarterUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quarter"

  # ─── NOTIFICATIONS ──────────────────────────────────────
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: Get notifications for current user
      parameters:
        - name: unreadOnly
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
  /notifications/{id}/read:
    post:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark a notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Marked read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
...
components:
  schemas:
...
    User:
      type: object
      required: [id, name, email, role, department, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [employee, manager, admin]
        department:
          type: string
        managerId:
          type: ["integer", "null"]
        managerName:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    UserInput:
      type: object
      required: [name, email, password, role, department]
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        role:
          type: string
          enum: [employee, manager, admin]
        department:
          type: string
        managerId:
          type: ["integer", "null"]
    UserUpdate:
      type: object
      properties:
        name:
          type: string
        role:
          type: string
          enum: [employee, manager, admin]
        department:
          type: string
        managerId:
          type: ["integer", "null"]

    # ── Goal ────────────────────────────────────────────
    Goal:
      type: object
      required: [id, employeeId, thrustArea, goalTitle, goalDescription, unitOfMeasurement, target, weightage, approvalStatus, lockStatus, createdAt]
      properties:
        id:
          type: integer
        employeeId:
          type: integer
        employeeName:
          type: ["string", "null"]
        thrustArea:
          type: string
        goalTitle:
          type: string
        goalDescription:
          type: string
        unitOfMeasurement:
          type: string
          enum: [numeric, percentage, timeline, zero_based]
        scoreType:
          type: string
          enum: [higher_is_better, lower_is_better]
        target:
          type: number
        deadline:
          type: ["string", "null"]
          format: date
        weightage:
          type: number
        achievement:
          type: ["number", "null"]
        progressScore:
          type: ["number", "null"]
        approvalStatus:
          type: string
          enum: [draft, pending, approved, rejected, returned]
        lockStatus:
          type: boolean
        quarter:
          type: ["string", "null"]
        quarterlyUpdates:
          type: array
          items:
            $ref: "#/components/schemas/QuarterlyUpdate"
        comments:
          type: array
          items:
            $ref: "#/components/schemas/Comment"
        isShared:
          type: boolean
        sharedGoalId:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: ["string", "null"]
          format: date-time
    GoalInput:
      type: object
      required: [thrustArea, goalTitle, goalDescription, unitOfMeasurement, target, weightage]
      properties:
        thrustArea:
          type: string
        goalTitle:
          type: string
        goalDescription:
          type: string
        unitOfMeasurement:
          type: string
          enum: [numeric, percentage, timeline, zero_based]
        scoreType:
          type: string
          enum: [higher_is_better, lower_is_better]
        target:
          type: number
        deadline:
          type: string
          format: date
        weightage:
          type: number
        quarter:
          type: string
    GoalUpdate:
      type: object
      properties:
        thrustArea:
          type: string
        goalTitle:
          type: string
        goalDescription:
          type: string
        unitOfMeasurement:
          type: string
          enum: [numeric, percentage, timeline, zero_based]
        scoreType:
          type: string
          enum: [higher_is_better, lower_is_better]
        target:
          type: number
        deadline:
          type: string
          format: date
        weightage:
          type: number
        achievement:
          type: number
...
    QuarterlyUpdate:
      type: object
      required: [quarter, achievement, progressStatus, updatedAt]
      properties:
        quarter:
          type: string
        achievement:
          type: number
        progressStatus:
          type: string
          enum: [not_started, on_track, completed, behind]
        progressScore:
          type: ["number", "null"]
        managerComment:
          type: ["string", "null"]
        updatedAt:
          type: string
          format: date-time
...
    Comment:
      type: object
      required: [id, authorId, authorName, role, text, createdAt]
      properties:
        id:
          type: integer
        authorId:
          type: integer
        authorName:
          type: string
        role:
          type: string
        text:
          type: string
        createdAt:
          type: string
          format: date-time
...
    SharedGoal:
      type: object
      required: [id, primaryOwnerId, goalTitle, goalDescription, target, unitOfMeasurement, linkedEmployeeIds, createdAt]
      properties:
        id:
          type: integer
        primaryOwnerId:
          type: integer
        primaryOwnerName:
          type: ["string", "null"]
        goalTitle:
          type: string
        goalDescription:
          type: string
        thrustArea:
          type: string
        target:
          type: number
        unitOfMeasurement:
          type: string
          enum: [numeric, percentage, timeline, zero_based]
        linkedEmployeeIds:
          type: array
          items:
            type: integer
        linkedEmployeeNames:
          type: array
          items:
            type: string
        achievement:
          type: ["number", "null"]
        quarter:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    SharedGoalInput:
      type: object
      required: [primaryOwnerId, goalTitle, goalDescription, target, unitOfMeasurement, linkedEmployeeIds]
      properties:
        primaryOwnerId:
          type: integer
        goalTitle:
          type: string
        goalDescription:
          type: string
        thrustArea:
          type: string
        target:
          type: number
        unitOfMeasurement:
          type: string
          enum: [numeric, percentage, timeline, zero_based]
        linkedEmployeeIds:
          type: array
          items:
            type: integer
        quarter:
          type: string
    SharedGoalUpdate:
      type: object
      properties:
        achievement:
          type: number
        linkedEmployeeIds:
          type: array
          items:
            type: integer

    # ── Quarter ─────────────────────────────────────────
    Quarter:
      type: object
      required: [id, name, year, startDate, endDate, isActive, isGoalSettingOpen]
      properties:
        id:
          type: integer
        name:
          type: string
        year:
          type: integer
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        isActive:
          type: boolean
        isGoalSettingOpen:
          type: boolean
        adminOverride:
          type: boolean
    QuarterInput:
      type: object
      required: [name, year, startDate, endDate]
      properties:
        name:
          type: string
        year:
          type: integer
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        isActive:
          type: boolean
        isGoalSettingOpen:
          type: boolean
    QuarterUpdate:
      type: object
      properties:
        isActive:
          type: boolean
        isGoalSettingOpen:
          type: boolean
        adminOverride:
          type: boolean

    # ── Notification ────────────────────────────────────
    Notification:
      type: object
      required: [id, userId, type, title, message, isRead, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        type:
          type: string
          enum: [goal_submitted, goal_approved, goal_rejected, goal_returned, quarterly_reminder, escalation, system]
        title:
          type: string
        message:
          type: string
        isRead:
          type: boolean
        relatedGoalId:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time

    # ── Audit Log ───────────────────────────────────────
...